Port chiaki-lib to Playstation Vita.
This is the first in (probably) a series of commits to port Chiaki to PS
Vita. The Vita uses a custom NetBSD-based network stack with
vendor-prefixed syscalls. There exists a port of libc, but I ran into
issues when using the wrapped version of the native epoll functions, so
this port adds lots and lots of ifdefs all over the place to use Sony's
network function instead of the default libc functions.