Port threading.c to PS Vita, includes breaking API change.
For Vita, the existing pthread wrappers are pretty hit and miss, or at
least I couldn't get any of them to properly work. So just like with the
networking stack, we reimplement the threading API with Vita-specific
APIs.

Unforunately this requires a breaking change for other platforms as
well, since the Vita API requires a mutex to be associated with a
condition variable. The Chiaki code already follows this pattern, but
does not associate the two during initialization, which is expected by
the Vita API. So the signature of `chiaki_cond_*` functions has changed
to associate the mutex during initialization and not pass it explicitely
during waiting.